Publish preparation and bug fixed.
Fixed most of bugs for final publish.
But met some problem about android studio manager or some kind of that problems, and couldn't even read what is the true problem and also hard to found solution.
Force me to delete editor and change to new editor from 2021.3.22f1 to 2022.3.31f1.